# api/docs/swagger.yaml
definitions:
...
  models.CreateStaffTarif:
    properties:
      amount_for_card:
        type: integer
      amount_for_cash:
        type: integer
      name:
        type: string
      tarif_type:
        type: string
    type: object
...
  models.Response:
    properties:
      data: {}
      description:
        type: string
      statusCode:
        type: integer
    type: object
...
  models.StaffTarif:
    properties:
      amount_for_carsd:
        type: integer
      amount_for_cash:
        type: integer
      created_at:
        type: string
      deleted_at:
        type: string
      id:
        type: string
      name:
        type: string
      tarif_type:
        type: string
      updated_at:
        type: string
    type: object
  models.StaffTarifResponse:
    properties:
      count:
        type: integer
      staff_tarifs:
        items:
          $ref: '#/definitions/models.StaffTarif'
        type: array
    type: object
...
  models.UpdateStaffTarif:
    properties:
      amount_for_carsd:
        type: integer
      amount_for_cash:
        type: integer
      id:
        type: string
      name:
        type: string
      tarif_type:
        type: string
    type: object
...
info:
  contact: {}
  description: This is a sample server celler server.
  title: Swagger Example API
  version: "1.0"
paths:
...
  /stafftarif:
    post:
      consumes:
      - application/json
      description: create a new stafftarif
      parameters:
      - description: staffTarif
        in: body
        name: staffTarif
        schema:
          $ref: '#/definitions/models.CreateStaffTarif'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.StaffTarif'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Create a new stafftarif
      tags:
      - stafftarif
  /stafftarif/{id}:
    delete:
      consumes:
      - application/json
      description: delete stafftarif
      parameters:
      - description: stafftarif_id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Delete stafftarif
      tags:
      - stafftarif
    get:
      consumes:
      - application/json
      description: get stafftarif by id
      parameters:
      - description: stafftarif_id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.StaffTarif'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Get stafftarif by id
      tags:
      - stafftarif
    put:
      consumes:
      - application/json
      description: get stafftarif
      parameters:
      - description: stafftarif_id
        in: path
        name: id
        required: true
        type: string
      - description: stafftarif
        in: body
        name: stafftarif
        schema:
          $ref: '#/definitions/models.UpdateStaffTarif'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.StaffTarif'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Update stafftarif
      tags:
      - stafftarif
  /stafftarifs:
    get:
      consumes:
      - application/json
      description: get stafftarif list
      parameters:
      - description: page
        in: query
        name: page
        type: string
      - description: limit
        in: query
        name: limit
        type: string
      - description: search
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.StaffTarifResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Get stafftarif list
      tags:
      - stafftarif
...
swagger: "2.0"